+++ shvcli/lsdir.py
@@ -1,21 +1,27 @@
 """Special handling of ls and dir methods."""
+import itertools
+import pathlib
 import string
 
-from shv import RpcMethodDesc, RpcMethodFlags
+from shv import Cpon, RpcError, RpcMethodDesc, RpcMethodFlags
 
 from .client import Node, SHVClient
 from .config import CliConfig
 from .parse import CliItems
-from .tools import print_flist
+from .tools import cpon_ftext, print_flist, print_row
 
 
-def ls_node_format(node: Node | None, name: str) -> tuple[str, str]:
+def ls_node_format(
+    name: str, node: Node | None = None, parent_node: Node | None = None
+) -> tuple[str, str]:
     """Print format for single node info."""
     nodestyle = "ansigray" if name.startswith(".") else ""
-    if node is not None and (subnode := node.get(name, None)) is not None:
-        if "get" in subnode.methods:
+    if node is None and parent_node is not None:
+        node = parent_node.get(name, None)
+    if node is not None:
+        if "get" in node.methods:
             nodestyle = "ansiyellow"
-        elif subnode.nodes:
+        elif node.nodes:
             nodestyle = "ansiblue"
     return (
         nodestyle,
@@ -44,11 +50,66 @@ def dir_method_format(method: RpcMethodDesc) -> tuple[str, str]:
 async def ls_method(shvclient: SHVClient, config: CliConfig, items: CliItems) -> None:
     """SHV ls method that is just smarter than regular call."""
     shvpath = items.interpret_param_path(config)
+    await shvclient.ls(shvpath)
     node = shvclient.tree.get_path(shvpath)
-    print_flist(ls_node_format(node, n) for n in await shvclient.ls(shvpath))
+    assert node is not None
+    if config.autoget:
+        for nn, nv in dict(node).items():
+            if not nv.methods_probed:
+                try:
+                    await shvclient.dir(str(pathlib.PurePosixPath(shvpath) / nn))
+                except RpcError:
+                    pass
+        if any("get" in nv.methods for nv in node.values()):
+            w = max(len(n) for n in node.keys())
+            for nn, nv in node.items():
+                n = [("", " " * (w - len(nn))), ls_node_format(nn, nv)]
+                if "get" in nv.methods:
+                    try:
+                        resp = await shvclient.call(
+                            str(pathlib.PurePosixPath(shvpath) / nn), "get"
+                        )
+                    except RpcError:
+                        pass
+                    else:
+                        print_row(
+                            itertools.chain(
+                                iter(n + [("", "  ")]), cpon_ftext(Cpon.pack(resp))
+                            )
+                        )
+                        continue
+                print_row(n)
+            return
+    print_flist(ls_node_format(nn, nv) for nn, nv in node.items())
 
 
 async def dir_method(shvclient: SHVClient, config: CliConfig, items: CliItems) -> None:
     """SHV dir method that is just smarter than regular call."""
     shvpath = items.interpret_param_path(config)
-    print_flist(dir_method_format(d) for d in await shvclient.dir(shvpath))
+    dirr = await shvclient.dir(shvpath)
+    if config.autoget and any(_use_autoget(d) for d in dirr):
+        w = max(len(d.name) for d in dirr)
+        for d in dirr:
+            n = [("", " " * (w - len(d.name))), dir_method_format(d)]
+            if _use_autoget(d):
+                try:
+                    resp = await shvclient.call(shvpath, d.name)
+                except RpcError:
+                    pass
+                else:
+                    print_row(
+                        itertools.chain(
+                            iter(n + [("", "  ")]), cpon_ftext(Cpon.pack(resp))
+                        )
+                    )
+                    continue
+            print_row(n)
+    else:
+        print_flist(dir_method_format(d) for d in dirr)
+
+
+def _use_autoget(method: RpcMethodDesc) -> bool:
+    return (
+        RpcMethodFlags.GETTER in method.flags
+        and RpcMethodFlags.LARGE_RESULT_HINT not in method.flags
+    )